requirements.txt
apprise
beautifulsoup4
boto3
PyYAML
redis
requests
tenacity
